Disease-Detection-From-Potato-Leaves-With-CNN
The primary objective of this deep learning project is to develop a robust model capable of accurately identifying and categorizing diseases in potato leaves. The target categories include Early Blight, Late Blight, and a Healthy state. Given a dataset of potato leaf images labeled with these categories, the model aims to learn distinctive features that differentiate between various states of potato health.